Add support for 'version_ttl'
Versions can have TTLs associated with them. When a version is past its
TTL it is archived, if it’s the latest version the key will not return
any data, only metadata. The version is however not fully deleted, just
marked as archived. A privileged user can still retrieve these versions
until they are fully removed by the max version check or are explicitly
deleted. Additionally an archived version can be unarchived by a user
with those permissions. Once unarchived a new TTL can be set, or the key
TTL will be restored. TTLs can be configured for the backend, a single
key, or a single version. The backend’s TTL will act as a default TTL
for all keys, the key’s TTL can overwrite the backend’s TTL, and the
version’s TTL can be less than the key’s TTL or the backend’s TTL.