Fix Renew operation for KV v1
Using `handleRead` as the `Renew` method for KV v1 is incorrect and
would fail with the error:

	http: panic serving 127.0.0.1:59579: field path not in the schema

in the Vault server without returning a response.

This fixes by returning an empty response which signals properly that
the secret cannot be renewed.

Also remove `GeneratesLeases()` which was never used and fix a typo in
an error message.