Remove unused code
The `Passthrough` interface ceased to be used after some code was
removed in #11.

The `GenerateLeases` function ceased to be used when this code was
made capable of being a separate plugin, rather than a builtin part of
Vault - credit to @remilapeyre for noticing this in #55 - I'm just
cherrypicking the removal of unused code from that old unmerged PR,
whilst I had some other unused code to PR the removal of too.