Port a few minor fixes over from the copy of this code in the Vault repo (#116)
* Port a few minor fixes over from the copy of this code in the Vault repo

There is a copy of this code in the Vault repo, to which some fixes have
been applied there but not here. Copy them over.

* Remove test change from submission

Upon reading through passthrough_test.go, I noticed that this was
a change in style of a single test function, to deviate from all the
rest. Therefore, I don't think it should be ported as is anymore.