Add `secret_path`, `operation` to metadata.
This preserves the existing `path` entry in the metadata for
backwards compatibility.

The `secret_path` is the path to the secret in Vault.

The `operation` is what operation is being performed, which may
be plugin-dependent.

This depends on https://github.com/hashicorp/vault/pull/22487
being merged and bumped in `go.mod` before this will compile
(since it references new constants in the SDK, though this could
be removed).