Fix panic on renewing a renewable KV v1 secret (#118)
* Fix panic on renewing a renewable KV v1 secret

A regression was introduced in #17, in 2018. Better late than never,
let's fix it. This will open the door to considering deleting the second
copy of the KV v1 source code within the Vault repo, and just referring
to this as the canonical copy.

The fix selectively reverts some parts of #17, renames `handleRead` to
`handleReadOrRenew`, and adds some detailed comments to explain the
subtlety.

* Oops, forgot to commit this file

* Add CHANGELOG entry

---------

Co-authored-by: robmonte <17119716+robmonte@users.noreply.github.com>